added the option to add new refrencevalues when picking one

+++ Assets/Scripts/HUD/ValueEntry.cs
@@ -15,6 +15,6 @@ public class ValueEntry : Entry
 
 	public override bool IsParentingType()
 	{
-		return Type > (int)ParameterType.POS;
+		return Type > ParameterType.POS;
 	}
 }
